requirements.txt
streamlit
torch
numpy
Pillow
opencv-python
PyYAML
matplotlib
requests
